+++ src/cmds/qalter.c
@@ -56,6 +56,7 @@
 #include "pbs_ifl.h"
 #include <pbs_config.h>   /* the master config generated by configure */
 #include <pbs_version.h>
+#include "portability.h"
 
 
 /**
@@ -293,9 +294,7 @@ main(int argc, char **argv, char **envp) /* qalter */
 					errflg++;
 					break;
 				}
-#ifdef WIN32
-				back2forward_slash2(optarg);
-#endif
+				fix_path(optarg, 2);
 				i = parse_equal_string(optarg, &keyword, &valuewd);
 				while (i == 1) {
 					set_attr_error_exit(&attrib, keyword, valuewd);
