+++ src/resmom/mom_main.c
@@ -1843,10 +1843,8 @@ usecp(char *value)
 	if (((pcphosts+cphosts_num)->cph_to = strdup(skipwhite(pnxt))) == NULL)
 		return HANDLER_FAIL;
 
-#ifdef	WIN32
-	back2forward_slash((pcphosts+cphosts_num)->cph_from);
-	back2forward_slash((pcphosts+cphosts_num)->cph_to);
-#endif
+	fix_path((pcphosts+cphosts_num)->cph_from, 1);
+	fix_path((pcphosts+cphosts_num)->cph_to, 1);
 
 	cphosts_num++;
 
@@ -7042,7 +7040,7 @@ main(int argc, char *argv[])
 				malloc(strlen(pbs_conf.pbs_home_path) + 17))) {
 				sprintf(pbs_conf.pbs_environment, "%s/pbs_environment",
 					pbs_conf.pbs_home_path);
-				back2forward_slash(pbs_conf.pbs_environment);
+				fix_path(pbs_conf.pbs_environment, 1);
 			}
 		}
 #else
